[slate] Remove settings and 2-monitor setup
In order to keep the file readable we remove settings that just reflect
default settings without customizing anything (we don't need to echo the
default settings).

We also remove 2-Monitor setup for now because this happens seldomly.
I first want to get comfortable with a single monitor coolness and
*understand* how all works rather than being overwhelmed by all the
settings.