Document `rcup` workflow, remove unneed instructions
We mainly cleanup the outdated content & simplify the initial `rcup` call

In order to keep the exclude string readable, we use explicit INCLUDE
for the first run.

First we create the `rcrc` config, then we install the rest with
settings from `.rcrc` file.

This allows us to configure all exclude pattern in one place.