sublime: Add settings & use official folder name
The goal is to ease the management of the preferences and installed
packages. The previous approach did not cover all settings.

Now we store all files in `Package` and `Installed Packages` inside
this repository and link these folders rather than only using the
`User` folder.

The symlinks are set within `hooks/post-up` and SublimeText should be
completely configured after applying the dotfiles.

*Developer Notes*

Another idea was to put the folder into `tag-ubuntu` and set `rscup`
set the links for us.

This approach didn't work out so well because `rcup` somehow got
confused and failed linking files.